tools/frontend_ui/requirements.txt
Flask==2.3.3
Jinja2==3.1.2
mysql-connector-python==8.1.0
python-dotenv==1.0.0